requirements.txt
# Core requirements for Evergrow360 backend
Flask==3.0.0
Werkzeug==3.0.1

# Firebase integration (simplified for demo)
google-cloud-firestore>=2.10.0

# Authentication & Security
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Limiter>=3.0.0
passlib==1.7.4
bcrypt>=4.0.0
cryptography>=40.0.0

# OpenAI integration (mock for demo)
openai>=1.0.0

# Payment processing (mock for demo)
stripe>=7.0.0

# Data validation
marshmallow>=3.20.0
email-validator>=2.0.0

# Environment management
python-dotenv>=1.0.0

# Production server
gunicorn>=21.0.0

# Development & testing
pytest>=7.0.0
pytest-flask>=1.3.0